Fix interfaces and code refactoring
