+++ node/interfaces/bodyRequest.ts
@@ -1,0 +1,4 @@
+export interface BodyRequest {
+  id: number
+  score: number
+}
