+++ node/interfaces/bodyResponse.ts
@@ -1,0 +1,7 @@
+export interface BodyResponse {
+  id: number
+  score: number
+  success: string
+  error?: number
+  errorMessage?: string
+}
