+++ node/interfaces/responseProduct.ts
@@ -1,0 +1,23 @@
+export interface ResponseProduct {
+  Id: number
+  Name: string
+  DepartmentId: number
+  CategoryId: number
+  BrandId: number
+  LinkId: string
+  RefId: string
+  IsVisible: boolean
+  Description: string
+  DescriptionShort: string
+  ReleaseDate: string
+  KeyWords: string
+  Title: string
+  IsActive: boolean
+  TaxCode: string
+  MetaTagDescription: string
+  SupplierId: number
+  ShowWithoutStock: boolean
+  AdWordsRemarketingCode: string
+  LomadeeCampaignCode: string
+  Score: number
+}
