+++ node/interfaces/scoreUpdateResponse.ts
@@ -1,0 +1,4 @@
+export interface ScoreUpdateResponse {
+  Id: number
+  Score: number
+}
