+++ node/middlewares/productScoreMiddleware.ts
@@ -1,3 +1,6 @@
+import type { BodyResponse, ResponseManager } from '../interfaces'
+import { buildErrorServiceResponse, buildResponse } from './utils'
+
 export async function productScoreMiddleware(
   ctx: Context,
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -15,9 +18,7 @@ export async function productScoreMiddleware(
     errors429: [],
   }
 
-  async function updateScore(updateRequest: BodyRequest): Promise<void> {
-    const { id, score } = updateRequest
-
+  async function updateScore(id: number, score: number): Promise<void> {
     try {
       const product = responseManager.responseProduct.find((p) => {
         return p.Id === id
@@ -62,13 +63,13 @@ export async function productScoreMiddleware(
     await retryCall()
   }
 
-  async function retryCall() {
-    const retryList: BodyRequest[] = []
-    let value = '0'
-
+  async function retryCall(): Promise<true | void> {
     if (responseManager.errors429.length >= 1) {
-      for (const index in responseManager.errors429) {
-        const response = responseManager.errors429[index]
+      const retryList = responseManager.errors429
+      let value = '0'
+
+      for (const index in retryList) {
+        const response = retryList[index]
 
         if (response.errorMessage && response.errorMessage > value) {
           value = response.errorMessage
@@ -77,15 +78,8 @@ export async function productScoreMiddleware(
         if (value === '0') {
           value = '20'
         }
-
-        retryList.push({
-          id: response.id,
-          score: response.score,
-        })
       }
-    }
 
-    if (retryList.length >= 1) {
       const awaitTimeout = (delay: string) =>
         new Promise((resolve) => setTimeout(resolve, parseFloat(delay) * 1000))
 
@@ -105,8 +99,10 @@ export async function productScoreMiddleware(
         responseManager.updateResponse
       )
       await Promise.all(
-        retryList.map(async (item) => {
-          return updateScore(item)
+        retryList.map(async (elem) => {
+          const { id, score } = elem
+
+          return updateScore(id, score)
         })
       )
 
@@ -118,37 +114,19 @@ export async function productScoreMiddleware(
 
   try {
     await Promise.all(
-      validatedBody.map(async (arg) => {
-        return updateScore(arg)
+      validatedBody.map(async (elem) => {
+        const { id, score } = elem
+
+        return updateScore(id, score)
       })
     )
     await myOperations()
 
-    const successfulResponses = responseManager.updateResponse.filter((e) => {
-      return e.success !== 'false'
-    })
-
-    const failedResponses = responseManager.updateResponse.filter((e) => {
-      return e.success === 'false'
-    })
-
-    ctx.status = 200
-    ctx.body = {
-      successfulResponses: {
-        elements: successfulResponses,
-        quantity: successfulResponses.length,
-      },
-      failedResponses: {
-        elements: failedResponses,
-        quantity: failedResponses.length,
-      },
-      total: responseManager.updateResponse.length,
-    }
+    buildResponse(responseManager, ctx)
 
     await next()
   } catch (error) {
-    ctx.status = 500
-    ctx.body = error
+    buildErrorServiceResponse(error, ctx)
     await next()
   }
 }
